Deprecate photon species list (#6265)
This PR deprecates `particles.photon_species`, since it is redundant (we
can use species type to define a photon species).
Specifically:
-  `particles.photon_species` is removed from the documentation
- `particles.photon_species` is removed from all the examples (in all
the examples we were already using `species_type`)
- the new logic in `MultiParticleContainer.cpp` is that a photon species
is initialized as such if `species_type` is set to `photon` or if it
listed in the `particles.photon_species`
- If `particles.photon_species` is used a warning is generated stating
that `photon_species` is deprecated
- An error message in `PhotonParticleContainer.cpp` is reworded to take
into account the preferred way of defining a photon species.
- Explicitly setting particles' mass to zero is now forbidden. This is
potentially a breaking change, but in my opinion it is very unlikely to
generate issues.