DSMC: implement generic two-product reaction (#6250)
**Description edited by @RemiLehe**

# Overview

This replaces the DSMC `CHARGE_EXCHANGE` process by a
`TWO_PRODUCT_REACTION` process, whereby:
- The reaction can have an associated energy cost (or gain) - while
`CHARGE_EXCHANGE` assumed no energy cost/gain.
- The reactant and products can have any mass/charge - while
`CHARGE_EXCHANGE` implicitly assumed that reactants and products had the
same masses, and imposed that their charge is incremented/decremented by
one during the reaction.

This change is backward compatible for users ; input scripts using
`charge_exchange` will work, but will internally call the new
`TWO_PRODUCT_REACTION` (with 0 energy cost/gain)

# Implementation

This leverages the function `TwoProductComputeProductMomenta` which is
also used in the fusion code. This function has been modified to add an
option to not scatter the products isotropically (which is used in DSMC)

As a result the calculation is valid for relativistic particles too.

# CI test

We added a test that is very similar to `test_1d_charge_exchange_dsmc`
(i.e. a beam of H+ goes through an H2 gas background, leading to the
charge exchange reaction `H+ + H2 -> H + H2+`).

However, while the beam energy in `test_1d_charge_exchange_dsmc` is 12
keV, the beam energy in the new test is only 5 eV. This means that the
energy consumed by the charge exchange reaction (1.66 eV in this case)
is not negligible, therefore justifying the need to use
`twoproduct_reaction` here. The new test checks the final energy of the
H and H2+ products, and the fact that it is consistent with the 1.66 eV
consumed by the reaction.

---------

Co-authored-by: Remi Lehe <remi.lehe@normalesup.org>
Co-authored-by: Roelof Groenewald <40245517+roelof-groenewald@users.noreply.github.com>
Co-authored-by: Edoardo Zoni <59625522+EZoni@users.noreply.github.com>