WarpXFaceExtensions: move some functions used only in the cpp file to anonymous namespace (#6309)
The main thing that this PR does is to move functions only used inside
`WarpXFaceExtensions.cpp` to an anonymous namespace in that file.
This PR also removes the `WarpXFaceExtensions.H` header, which was
already unused. The dosctrings of the function definitions are moved to
the cpp file.
~~Finally, `amrex::Abort` is replaced with `WARPX_ABORT_WITH_MESSAGE`
for consistency.~~ <-- can't do that because it's in device code