Generalize DSMC two-product reactions for photons (#6290)
**Description edited by @RemiLehe**

This generalizes the function `TwoProductComputeProductMomenta` so that
it can be used for photons. (This enables the possibility to model
photo-ionization and recombination with DSMC.)

Instead of using the normalized momentum `u = p/m` (which is not defined
for photons), the operations in `TwoProductComputeProductMomenta` have
been rewritten to use `p` directly.

In addition, this PR adds:

- [x] CI test for photo-neutralization reaction: `Hminus + photons ->
Hneutral + electrons`
- [x] analysis script to check product species velocities (simulation
vs. theory).

---------

Co-authored-by: Remi Lehe <remi.lehe@normalesup.org>