Rework the PECInsulator boundary conditions (#6253)
NOTE: This PR is not complete and still undergoing testing.

This reworks the PECInsulator boundary conditions to be more correct,
specifically to fix conservation of energy. In the simplest form, the
conditions for both PEC and insulator (on the upper x axis at L) are
```
E||(x > L) = 2*E||(L) - E||(2*L - x)
Ep(x > L) = Ep(2*L - x)
B||(x > L) = B||(2*L - x)
Bp(x > L) = 2*Bp(L) - Bp(2*L - x)
```
Where `||` and `p` are parallel and perpendicular to the boundary.
However, the actual implementation is not so simple. For instance,
`E||(L)` and `Bp(L)` are zero for PEC, but not for insulator. Also,
depending on whether `E` or `B` is specified in the insulator, either
`E||(L)` or `B||(L)` will be specified or be evolved (and so unchanged
here) which means that in the extrapolations either the evolved value or
the specified values needs to be used.