Virtual photons  (#5595)
This PR adds the option of creating a virtual photon population
associated to a given electron or positron species.
The virtual photons are generated from scratch at every timestep.
They are initialized in the same position as the parent particle.
Their momenta is sampled from the theoretical analytical spectrum (see
equation 99.16 in Berestetskii, Pitaevskii, Lifshitz. Quantum
Electrodynamics: Volume 4.)

The virtual photons can be used in linear Breit-Wheeler or linear
Compton collisions. This is used to simulate incoherent beam-beam
effects at the interaction point of colliders.

To do:
- [x] create a virtual photon species with the correct number of
particles and with attributes same as primary particles
- [x] draw the virtual photon energy from the correct probability
distribution
- [x] add infrastructure
- [x] add docs and test

Note: As of this PR, the virtual photons are technically not virtual at
all, as their virtuality is zero. A more accurate name would be
"equivalent quanta". However, follow-up PRs will soon add virtuality
calculations.

🧨

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Luca Fedeli <luca.fedeli.88@gmail.com>
Co-authored-by: Edoardo Zoni <ezoni@lbl.gov>
Co-authored-by: Edoardo Zoni <59625522+EZoni@users.noreply.github.com>
Co-authored-by: Neïl Zaim <49716072+NeilZaim@users.noreply.github.com>
Co-authored-by: Andrew Myers <atmyers@lbl.gov>
Co-authored-by: Remi Lehe <remi.lehe@normalesup.org>